[571287] Update tests to correctly managed OffscreenEditPart
Currently the lifecycle of the OffscreenEditPart (and associated
EditPartViewer, Shell, Listeners, ...) is not correctly managed in tests
of MappingBasedSiriusFormatManagerFactory.
The Shell is disposed too early. As consequence, when the postCommit
listeners are called, the EditParts, especially DEdgeEditParts, are no
longer recorded as listeners. So they are not notified, and the effect
on Draw2D edges are not done. The final consequence is that the Draw2D
points list recorded for edges in tests were wrong. As we compared with
an expected result, generated from the tests themselves, we had not seen
this problem before.

The steps are now similar to:
* Get the offscreen edit part (by creating a new Shell)
* Execute the RecordingCommand (the post commit listeners are called at
the commit of the transaction)
* Deal with the offscreen edit part (mainly store the format data)
* Undo the command (to let the session, more or less, in its previous
state)
* Clean and dispose the offscreen edit part (and its associated Shell)

This commit contains also a change in
MappingBasedSiriusFormatManagerFactory: only to create one shell per
OffscreenEditPart. In reality, there is no impact with this modification
because there is only one Diagram for each DRepresentation. But the code
is now cleaner.

Bug: 571287
Cherry-picked-from: 571249
Change-Id: I74a706c5b1017d8688cc7327df30ca8f0d5af844
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>