[572073] Call revalidate() on a constraint change on label on border
If the constraint is changed, the figure must be "revalidate". This
problem has not been seen before because generally there are other
modifications/notifications that trigger the revalidation. At least with
ELK, it is possible to have a change concerning only the label location
(and not on its parent). So a revalidation is necessary to be sure that
the figure is refreshed.

Bug: 572073
Change-Id: I0f3be5c6f0c8432662a12ca7c140174793b02f19
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>