[572073] Add a test to reveal the problem
This commit adds a new diagram to reveal the problem. In this diagram,
the arrange all only moves the label that is on a border. With this
commit, the expected result is to have a failure. Only the GMF
coordinates are OK, the Draw2D coordinates are wrong. This is the bug.

Bug: 572073
Change-Id: I42969248b30d16b07d34e7facb5acf734ee3e293
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>