[571925] Use options Priority and Node Placement Strategy
This commit adds a bigger "Priority" [1] on edges used as source or
target to give preferential treatment for this edge (more flat than
edges connected to it). This option must be used in correlation with
"NETWORK_SIMPLEX" as "Node Placement Strategy" [2]. Otherwise, the
"Priority" is ignored.
As conclusion, the result is better with this commit.

[1] https://www.eclipse.org/elk/reference/options/org-eclipse-elk-priority_org-eclipse-elk-layered.html
[2] https://www.eclipse.org/elk/reference/options/org-eclipse-elk-layered-nodePlacement-strategy.html

Bug: 571925
Change-Id: I8864fff9661570e15cfb838c4a8537d40bc9947b
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>