[571056] Fix STD004 test
The test org.eclipse.sirius.tests.swtbot.std.STD004.testSTD004() fails
since the addition of test
org.eclipse.sirius.tests.swtbot.propertypage.SiriusSessionDetailsPropertyPageTest.testSiriusSessionDetailsPage()
in the SWTBot tests suite in commit 36107572 [1]. This test closes the
Outline view in the setUp methods but does not reopen it in the
tearDown. As the STD004 needs the Outline view, it fails with a
WidgetNotFoundException: Could not find view matching: with title
'"Outline"'.

This commit just reopens the Outline view in the tearDown(). A
"SWTBotUtils.waitAllUiEvents();" is added like in other similar tests.

[1] https://git.eclipse.org/c/sirius/org.eclipse.sirius.git/commit/?id=36107572fe9d3f132fe342642a3c12e04f4dedd9

Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=571056
Change-Id: I717416e2e52e8b8750787bcd8e9a46d661771654
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>