[test] Fix some JUnit 4 tests
This commit fixes some JUnit 4 tests: Some tests have not a public
"tearDown()" method with "@After" annotation. As consequence, the
tearDown is not called. The environment is not cleaned for the following
tests. For example, with all tests inheriting from
AbstractMappingBasedSiriusFormatDataManagerTest, the following tests are
executed with "refreshAtOpening" equals to false instead of true. This
is problematic for example for TablesAndEntitiesDirtyTest on a CI
server.

This commit also adds some "wait" in TablesAndEntitiesDirtyTest as this
test is the initial focus.

Change-Id: Ie2c4398f79a4f986fd2e9cb503a4d8839c4bccf0
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>