[572638] Mutualize code to warn not managed Object in LayoutDataManager
This commit mutualizes the code to warn the user that some kind of
objects are not managed by the LayoutDataManager. The main goal of this
commit is to avoid to log an unexpected warning for DNodeListElement.
This is already the case for AbstractSiriusFormatDataManager but not for
LayoutData and RootLayoutData.

FormatData_illegalTarget message has also been removed. This is dead
code as FormatData class no longer exist.

Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=572638
Change-Id: I737f54b5872f7bf205eaccbaf582d3834253cf71
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>