[570518] Add migration participant to repair Edges with one Bendpoint
In some cases, it is possible to have only one Bendpoint on Rectilinear
Edges, which makes them invalid and it is impossible to move the
elements they connect to.
The added migration participant identifies these Edges that have only
one Bendpoint and repairs them by setting a minimum of two Bendpoints.

Bug: 570518
Change-Id: Ie646ad2349b0ea6f3caa46a649eff2dae436f5d4
Signed-off-by: Glenn Plouhinec <glenn.plouhinec@obeo.fr>