[test] Reduce length of a long test for SiriusFormatDataManager
The existing condition has been moved to reduce directly the data
instead of not doing the "core" of the test for some tests.
With this commit there are "only" 500 tests executed instead of 900.
Before this commit 400 tests are executed to do "nothing" but the setup
and the teardown are done whatever happens.

Change-Id: I3001e15236684cf662d3a523749286c27d68ada0
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>