[573086] Move rep desc validation from adapter to post commit listener
DRepresentationDescriptorQuery delegates the computation and storage of
the validity of a DRepresentationDescriptor to an adapter. However, when
undoing the deletion of a representation, this adapter was triggered too
early. At this point the DRepresentationDescriptor was not only
transient but was not yet reattached to the resource.
Now, the validation is triggered by a post commit listener instead.

Bug: 573086
Change-Id: I39c71de29a006ec60e4d28258c6d2aa07495dacd
Signed-off-by: Steve Monnier <steve.monnier@obeo.fr>