[571115] Avoid "magic" behaviors if a CellEditor tool is provided
For "generic" cell editor, ie provided directly by Sirius according to
feature type and characteristics, there are some "conversions" of the
value (in getter and setter). These "conversions" are not necessary and
can be disturbing for the specifier when it provides a CellEditor tool.
This commit avoids them when CellEditor tool is provided.

The documentation and tests are also adapted to reflect the new behavior
(no edition fallback).

Bug: 571115
Change-Id: I0c151f322e93e9464e2dac0e574ae47da50adf97
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>
Signed-off-by: Nicolas PERANSIN <nicolas.peransin@obeo.fr>