[574241] Wrong Note text alignment in collaborative environment
When a Note is created in a Sirius diagram in a collaborative context
(Obeo Designer Team Edition or Team for Capella), the text is centered
vertically (instead of being aligned at the top).
Indeed, the implementation of ESTRING_TO_STRING_MAP_ENTRY is not
obligatorily a org.eclipse.emf.ecore.impl.EStringToStringMapEntryImpl.
So this commit adapts the code of
org.eclipse.sirius.diagram.ui.internal.view.factories.SiriusNoteViewFactory.createDefaultVerticalAlignmentEAnnotation()
to handle this case.
This commit also adds a migration participant to fix the notes created
before this version.

Tests:
* org.eclipse.sirius.tests.unit.diagram.migration.NoteShapeDefaultLabelAlignmentMigrationTest
always checks the migration of initial problem (bugzilla 515044).
* org.eclipse.sirius.tests.unit.diagram.migration.NoteShapeDefaultLabelAlignmentMigrationTest2
checks the migration of the new problem corresponding to this bugzilla.
* org.eclipse.sirius.tests.swtbot.NoteCreationTest has been updated to
also check that the GMF styles concerning note text alignment.

Bug: 574241
Cherry-picked-from: 574133
Change-Id: I2acff0a2f69d9623e1576d46759b521b01c0bec9
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>