[574273] Add z-order actions on edges
* The actions are provided with Eclipse menu extension points (to allow
shortcuts easily).
* Actions are "disabled" for sequence diagram.

For information, even if actions are only for edges, the API added in
IDiagramCommandFactory also handle the nodes cases (code copied from
GMF). This behavior is not used currently but maybe in another
improvement to homogenize the behavior between nodes and edges (shortcut
for example).

Some explanations about the changes of this commit:
* All classes in org.eclipse.sirius.diagram.tools.api.command are the
code for the commands finally executed and who change the model.
* The changes in sequence.ui plugin are only little changes to have no
effect on sequence diagrams with the new actions.
* The changes in
org.eclipse.sirius.diagram.ui.tools.internal.actions.layout and
org.eclipse.sirius.diagram.ui.tools.internal.commands are to manage the
actions available in contextual menus.
* The changes in
org.eclipse.sirius.diagram.ui.edit.api.part.AbstractDDiagramEditPart are
to reorder the edges figures according to GMF order (to reflect the
changes done by the actions/commands on the figure).

The change in
org.eclipse.sirius.diagram.ui.edit.api.part.AbstractDDiagramEditPart.refreshChildren()
has probably a "performance" impact because we redraw the edge figure in
a "new order". This performance is "not significant" because this method
is called in specific conditions:
* org.eclipse.sirius.diagram.ui.edit.api.part.AbstractDDiagramEditPart.handleNotificationEvent(Notification)
calls refreshChildren() when
    * VIEW__PERSISTED_CHILDREN or VIEW__TRANSIENT_CHILDREN is changed
    * a NoteAttachment is removed
    * activatedFilters is changed
    * activatedLayers or activatedTransientLayers is changed
    * when the diagram is notified of a set, a unset or an add
notification
    * on a visibility change on a GMF view
    * on a change of order of GMF edges
* after these actions
    * creation of a new element at the root of the diagram (edge, node,
container, note, note attachment)
	* external save on VSM or semantic resource (reload of the session and
refresh of the diagram)
* not on these actions
    * Move of a node/container
	* Move of an edge
	* Reconnect of an edge
	* Creation of a new element inside a container

Bug: http://eclip.se/tmpolicH
Change-Id: I83547fc2ac9fe2186c23271792cd8e1879eecd98
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>