[test] Remove overly specific test assertion
It fails in modern modulats JVMs with:

   Unable to make field final java.util.Collection
   java.util.Collections$UnmodifiableCollection.c accessible: module
   java.base does not "opens java.util" to unnamed module @53398bc6

Being this specific about the implementation class we expect of the
result does not make much sense anyway. If the properties we try to
test here (no duplicate, well-defined order) were part of the contract
of getSelectecViewpoints, then it should return a SortedSet instead of
a generic Collection.

Change-Id: I4a2d842c5a1154a9f9a4843a3db06688191fbce9
Signed-off-by: Pierre-Charles David <pierre-charles.david@obeo.fr>