[572899] Add additional Batik components in target platforms
Some tools built on top of Sirius includes Sirius .targetplatform
definitions without the use of "with requirements".

Adding 	org.apache.batik.bridge and 	org.apache.batik.script solves the
issue in the build of those components but this might impact other
products so a solution is to list them in the orbit location of Sirius
tps.

Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=572899
Change-Id: If340504803c5b5acfd26197df333d9c8afcc37e5
Signed-off-by: Maxime Porhel <maxime.porhel@obeo.fr>