[531282] Optimize mappings checks in DDS::processEdgeMappingsRefresh
This change has no impact on DiagramDescription with no EdgeMapping with
EdgeMapping references as source/target mapping.

When a diagram description contains a EdgeMapping with another
EdgeMapping as source or target mapping, there is no need to compute the
full structure of edge, node and container mappings each time we want to
test that the referenced mapping belongs to the current description. The
predicate has been move to the calling method as those information is
already available. It is now passed as parameter and only edge mappings
checks are done.

Bug: 531282
Change-Id: I28ab96dbdaa0e1b5137107829dae01d34446b288
Signed-off-by: Maxime Porhel <maxime.porhel@obeo.fr>