[575196] Enhance usability of the ImageSelectionDialog API
In order to override the ImageSelectionDialog in other plugins, the
refresh job family has been defined as a constant to manage the job more
easily, the ImageFiltersUtils utility class has also been moved to the
API package. Selection of elements in Gallery has been improved by
creating the GallerySelectable class for testing purposes.

Bug: 575196
Change-Id: Icea5d4a71359cc71926bd9714179f7b274c64571
Signed-off-by: Glenn Plouhinec <glenn.plouhinec@obeo.fr>