[fix] Prevent NPE when trying to get the representation
In some particular condition, when the Sirius session is a collaborative
session based on EMF CDO, DelegatingSessionProtocol.getSession throws a
NPE when trying to load the object revision.
It corresponds to a case when the session has been closed and the
CDORevisionManagerImpl has been disposed.
It that case, the exception is silently catched and no representation is
returned.

Cherry-picked-from: 6c6a6b436e7b6c4c18e52bdc4a8db4191831db56
Change-Id: Idfd15f5267e82336ffdec9a66732210d62bf453e
Signed-off-by: Laurent Fasani <laurent.fasani@obeo.fr>