[575196] Manage image persistency for EAttribute containing html
* Add a pre-commit listener to manage base64 image. The listener will
transform the base64 string to an image and update the path
consequently.
The image creation, the undo and the redo are delegated thanks to the
ImageManagement new API.
* ImageManagement implementation is provided by ImageManagerProvider
* RichTextAttributeRegistry has been added so that each modeler can
register the EAttribute that may contain base64 string.
* ImageManagement implementation also contains behavior to convert
string from/to html so that it can be properly displayed in a rich text
editor(html)

Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=575196
Change-Id: Ied49c8778d2d5d5bcc76f94af263b1dad16d1358
Signed-off-by: Laurent Fasani <laurent.fasani@obeo.fr>