[576057] Add migration participant to keep same edges display order
In some cases, see issue for details, the display order of edges is
different after the z-order feature implementation.
The goal of this migration participant is to change the order of GMF
edges to keep the same display order.

The version of aird used in ZOrderActionsTest has been changed. Indeed,
this data already contains specific edges order.

Bug: 576057
Change-Id: Ibb8e7663732fc2562a96d75861c06f8522a14533
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>