[576722] Fix z-order regression
Since the z-order (bugzilla 574273), there is a potential regression
that occurs in a context of "unsynchronized diagrams".
It seems possible that the method
AbstractDDiagramEditPart.refreshChildren() is called with some edges
that are visible at the GMF level but without a corresponding edit part.
I think that it is a temporary state but this method must handle it.

A corresponding test will be added later, when the causes will be better
understood.

Bug: 576722
Change-Id: I01dedc16aec1eb353b481020a3d35fe0bed2081c
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>