[577162] Hide/reveal selected label for edges
Update metamodel to add an Integer list attribute to HideLabelFilter in
order to identify which label of an edge should be hidden. These labels
corresponds to the VisualID of the specific edge label edit part.
There is also a first step of updating the command builders in order to
have the hide label functionality working on diagram from context menu.

Bug: 577162
Change-Id: I069356cf2153e52cd752ff586d0cdfdcf59164be
Signed-off-by: Steve Monnier <steve.monnier@obeo.fr>