[577045] Allow resize and move of an inner fragment
- Allow move/resize of a frames when future parents list contains one
operand among several parents.
- Continue to forbid resize/move when several operands are detected in
future parents: a Frame cannot have several Frames as parents. This
means that the moved frame would be above the two parents, ie. after the
moved frame would not be fully included in one of the parent frame.
- Remove check
- Move coverage computation as early step of the validator (even if it
can be cached in CacheHelper.getCoverageCache(), avoid to compute it
several times)
- Add NPE guard in SubEventHelper: NPE can be triggered in late/async
evaluations

Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=577045
Change-Id: I152dd3d09afdf84ca19446ee0c9e4c85a5e61f01
Signed-off-by: Maxime Porhel <maxime.porhel@obeo.fr>