[577397] Allow SVGFigure Transcoder override
In order to override the SVG stored in SVGFigure.ImageCache to apply some
image transformation prior to storing the image in the cache, one would
like to simply provide an overridden implementation of the render method
in SimpleImageTranscoder.

In order to do so this patch moves the transcoder creation in a
dedicated protected method. So any SVGFigure can provide its own
SimpleImageTranscoder and apply overriden render content.

Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=577397