[577676] Add specific unnecessarily calls for sequence diagram
As for previous commit, this new one allows to avoid to launch an
arrange all after a "sequence arrange" (RefreshLayoutCommand). It is not
necessary.

This commit adds a new API to add new "restrictions" concerning the
DOperationHistoryListener that launches arrange all after "Done"
operations. This new API allows to add new "restrictions" for specific
IDiagramDescriptionProvider (specifically only for sequence diagrams).

Bug: 577676
Change-Id: I2d6754c48566e6cdb7763696f5258ca2f63bc48b
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>