[577676] Same behavior for all containers on arrange all
At the creation of a diagram, the behavior was not the same for the
first container (that remain at same location; 0,0), than the others.

With this commit the behavior is now the same for all containers but the
draw2d size is always wrong for list elements with an initial size. The
size is reduced to fit the content instead of using the size defined in
the VSM. Another commit is necessary.

Bug: 577676
Change-Id: I4e8578c7ecaa4b13443dc8e840d652a40a53e41e
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>