[577676] Change behavior for regions during refresh bounds
The modified code in
org.eclipse.sirius.diagram.ui.edit.api.part.AbstractDiagramElementContainerEditPart.getDefaultDimension(),
used in refreshBounds(), has been added for region collapse cases.
But the case with a defined size for the regions has not been handled.
If the region is not collapsed, the default size, defined in the VSM,
must be restored.

org.eclipse.sirius.tests.swtbot.compartment.CompartmentsTest.testCollapseOfNewVerticalRegionContainerWithTwoRegions()
is a sample of test that was KO without this specific code.

This commit also added collapse tests for regions with a fixed size to
have a more complete code coverage.

Bug: 577676
Change-Id: I5000718c1f5a2fbaf239ee2666aeb1235b8486c3
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>