[573405] Make same size operational on regions container
This commit fixes and improves support of "make same size" action
applied on regions container:
* Action is now supported when regions container is used as source or as
target.
* Action is now supported when the selection contains both regions
containers and standard containers.
* For the already supported case, the regions container was not set as
auto-size, with {-1;-1}. The display size, ie draw2d size, is computed
later according to regions size.

The code coverage of SizeBothAction is 99% with the tests
CompartmentsSizeTest (updated in previous commits)

No change on documentation, it is already OK (see "Make Same Size" in
https://www.eclipse.org/sirius/doc/user/diagrams/Diagrams.html#ref_tabbar_element).

Bug: 573405
Change-Id: I59aa0675f9d1d11b2edd54872298601c1828983d
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>