[573405] Update CompartmentsSizeTest to reveal problem
The data has been modified to reveal the problem.

Currently a regions container should have {-1, -1} as size. But some
models, probably because of previous bug, can contain a real width or a
real height. The problem of this issue occurs only on this kind of
model. This is why the tests of CompartmentsSizeTest were OK before.

This commit also cleans up the code of CompartmentsSizeTest.

Bug: 573405
Change-Id: I35761d0958304cd6093eeebe6b2ad00ea3f80be1
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>