[576305] Improve draw2d refresh for regions containers
This commit handles the specific cases of "a region size" impacted by
the size its regions container (with VStack), itself impacted by a
"brother region", ie the regions container is itself a region and its
width is impacted by its brother.
For that, this commit adds a specific listener between the figure of the
region and its "parent figure" (not the direct parent, but the one with
the width change that must have effect).
There is maybe similar problem concerning height with HStack but it is
not the current issue.

The main changes are in AbstractDNodeContainerCompartmentEditPart. The
listener is added between the "container figure" and the "region
figure". The changes are focus on the "Capella Title Block" case. This
commit improves this case but as reminder, assembling multiple region
containers is not supported. And this commit is certainly not enough to
handle all assembly cases.


Bug: 576305
Change-Id: I97ac7052e962f1a14705b294118dd326ca62349a
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>