[577676] Fix tests regressions
This commit fixes tests KO since the commits about this issue:
* ArrangeAllWithSnapToGridTest.testBorderNodeSnapWithArrange: The 2
successive "Arrange All" was moving "LF1", but it was not really
expected. Now "LF1" stay on x axis to the waiting condition has been a
little modified.
* ContainerDefaultSizeLayoutTest.testDefaultSizeWithExternalCreationOnShape:
{-1, -1} is now the expected GMF bounds (because preference
PREF_AUTOSIZE_ON_ARRANGE is true by default). Before it was not the case
because of a bug for first element (see "[577676] Same behavior for all
containers on arrange all"). I also added another test to cover both
preference value.

Bug: 577676
Change-Id: I5e196792bd45676bb0b4c1b0eb5c03733443bff2
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>