[578321] fix detection broken repDesc.getRepPath() fragment
DRepresentationDescriptorQuery.isRepresentationReachable() did not load
representation to avoid the package explorer to load every
representation. Now the API have a boolean to force that loading so this
query can have a valid result for tests or the representation validation
tool.
The test raised some unexpected NPE in CommandStackUtil, so it was fixed
as well.

Bug: 578321
Change-Id: I103404bbc1c400a9a75f8bccfab6c44902bc72b5
Signed-off-by: Steve Monnier <steve.monnier@obeo.fr>