[576423] Update migration participant for WorkspaceImage
The migration participant added in previous commit was too intrusive. It
changes GMF nodes size even when it is not expected.
This commit reduces the scope of the impacted nodes to only consider the
ones corresponding to the bug.
As no release exists with this migration participant we can only update
it (without creating a new one).

The labels of some containers of tests data, for
WorkspaceImageGMFBoundsMigrationParticipantTest, have been reduced to
avoid problem of "Font size". This test has also been adapted to reflect
the above changes.

Bug: 576423
Change-Id: I32f0dcee6ba55d5cbc08d32342049461085e99ef
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>