[578436] Preserve selection during d'n'd or reconnect
This commit avoids to have 2 selections changed during a drag'n'drop of
an element (SiriusContainerDropPolicy) or during a reconnection of an
edge (SiriusGraphicalNodeEditPolicy).
These 2 selection changes were :
* one for the diagram selection, when the old edit part is removed,
* and one for the "same" element, when the new edit part is created. It
is not really the "same", but the underlying element is the same
(DDiagramElement and semantic element).

With this commit, the notification of the selection changes is disabled
during the drag'n'drop (or the reconnect). It is re-enabled through the
DiagramSelectDRepresentationElementsListener (after the drag'n'drop is
done and just before the "new selection").

Bug: 578436
Change-Id: I4153ed576e9cc534c9beb0c96331bf1e1791e535
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>