[579818] Ignore DDiagram_ActivatedFilters notif for other diagrams
Assuming diagram1 and diagram2 are opened.
When applying the filter on diagram1
* nothing is done for this notification for diagram2
* the filter application will generate some visibility attribute changes
in diagram1. We consider for this commit that some filters in diagram2
could be based on visibility changes of elements in diagram1.

So if n diagrams are opened we move from 2n calls of
computeCompositeFilterApplications to n+1 calls.

Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=579818
Change-Id: I4d1998f1f50751e1c36bf95d6655d2e7888fc3e2
Signed-off-by: Laurent Fasani <laurent.fasani@obeo.fr>