[579856] Properly initialize fill, stroke and stroke-width for used SVG
The Sirius export SVG, as it is done, do not structure the svg as it
could be with something related to the EditPart/Figure structure.

If a GMF Note is used, the embedded svg inherits the Note attributes.
This commit adds default fill, stroke and stroke-width values
(respectively black, black and 0) in the use tag.

Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=579856
Change-Id: Ia4d1b6de4a94af163814b33b4c3bb90b910be311
Signed-off-by: Laurent Fasani <laurent.fasani@obeo.fr>