[577162] Clean semanticToLabelsVisualIDToHideMap during dispose
The field semanticToLabelsVisualIDToHideMap has been added in commit
a1b0cfa9 [1] but it is not "disposed". In bugzilla 431634 [2], some work
has been done to clean the actions of tabbar, so this field must be
"disposed" as other. The test
org.eclipse.sirius.tests.swtbot.tabbar.TabBarTest.testTabbarDisposal()
fails since commit [1].

[1] https://git.eclipse.org/c/sirius/org.eclipse.sirius.git/commit/?id=a1b0cfa932975bb6a6f53d9e8039a4beef94f4be
[2] https://bugs.eclipse.org/bugs/show_bug.cgi?id=431634

Bug: 577162
Change-Id: I8f2636397b7139151f3bd576ad5e9c8a93444d53
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>