[test] Fix EdgeWithBorderNodeCreationPositionWithSnapToGridTest
This commit fixes the test
EdgeWithBorderNodeCreationPositionWithSnapToGridTest.testBorderNodesAreAlignedWithScrollOnDiagramAndWithZoom
and probably some other random failing tests. Indeed, the analysis
allows to detect that sometimes the reveal is not finished when the rest
of the test is executed. In this case, the coordinates used in the rest
of the test are wrong, and the failure is "normal".

This commit adds a wait condition at the end of reveal method to be sure
that the reveal is graphically finished.
It also reverts previous "try of fix".

Change-Id: I0f3c9b47c621ac4b1d9a3f342943a1ef12fb4a82
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>