[476523] Add edgeView and otherEnd in the reconnect precondition
This commit adds "edgeView" and "otherEnd" variables in the context of
the evaluation of the precondition of reconnect tool. Before, there were
available only during the tool execution.
This commit also adds completion for these variables in the VSM editor.

validateVariableTypes.odesign has been modified to be sure that the
variables edgeView and otherEnd are correctly validated in the VSM, with
the test org.eclipse.sirius.tests.unit.diagram.vsm.VSMVariableTypesValidationTest.

The test org.eclipse.sirius.tests.unit.api.tools.ReconnectionTest.testReconnectPreconditionEvaluation()
has been added to validate that the evaluation of the precondition is OK
when containing variables edgeView and otherEnd (by using the new
TaskHelper API).

The test org.eclipse.sirius.tests.swtbot.EdgeReconnectionTests.testSimpleEdgeSourceReconnectionWithObliqueStyleRouting()
has been completed to add a precondition with the new variables and
to check a OK case and a KO case.

Bug: 476523
Change-Id: Ia5c6656634d05666a2f00a8c5ffca59a0d3e81cb
Signed-off-by: Maxime Porhel <maxime.porhel@obeo.fr>
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>