[580225] Use only one label.setVisible during edge layout()
In the previous fix for this issue, commit d34edbb6 [1], two setVisible
method can be called in some cases. This seems to cause "draw2d"
regression in case of two edges creation. The test, in collaborative
environment,
fr.obeo.dsl.viewpoint.tests.collab.swtbot.lock.ImplicitLockDecorationOnSharedDiagramTests.testLockDecorationOnMiddleOfEdge
allows to detect this regression.

In this commit only one setVisible is used and now the behavior for edge
layout is the same than before the fix [1].

[1] https://git.eclipse.org/c/sirius/org.eclipse.sirius.git/commit/?id=d34edbb6f6f732f4b26f34a03c89c21175375a2c

Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=580225
Change-Id: I293a035e202ca6132b7ced5d31d40df1ef3bb101
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>