[15] Add automatic tests about FilterListener notifications changes
The data of this existing test class has been completed with 2 other
diagrams and an additional filter.

The comparator of FilterService has been changed in this commit:
org.eclipse.sirius.diagram.business.api.helper.filter.FilterService.sortFilters(...).new
Comparator() {...}.compare(FilterDescription, FilterDescription).
Indeed, the result is random. It was stable in a same runtime but it was
random between two successive runtimes. So the number of notifications,
calling FiltersSortingCommand in
DiagramListenersTests.testNbNotificationsWhenActivatingFilter(), was not
stable.

The test "testNbNotificationsWhenActivatingFilter()" reveals that we
have not the expected result of commit [1], "Ignore
DDiagram_ActivatedFilters notif for other diagrams". Indeed, the nb
calls to computeCompositeFilterApplications is always 2n. The only
changes is that FiltersSortingCommand is called only for the diagram
where the filter activation is done. This problem will probably be
handle later (for example with gerrit [2]).

[1] https://git.eclipse.org/c/sirius/org.eclipse.sirius.git/commit/?id=c94261ec2535c3f9d5f09c600edfb1c7a63c5036
[2] https://git.eclipse.org/r/c/sirius/org.eclipse.sirius/+/193088

Bug: https://github.com/eclipse-sirius/sirius-desktop/issues/15
Cherry-picked_from: https://bugs.eclipse.org/bugs/show_bug.cgi?id=579818
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>