[580524] Support project dependencies without opened Session
* The image project dependencies are persisted in a DAnnotationEntry on
the main DAnalysis. A precommit listener has been created to add and
remove dependencies to images projects for each modification of the
WorkspaceImage style.
* A new class SiriusProjectDependencyQuery is added to get serialized
project dependencies from a given project. It is based on registered
IProjectDependencyProvider. This commit provides one implementation for
local project that uses SAXParser to leverage the serialized
information.
* SessionDetailsReport is updated to always show the general
dependencies (aird, semantic resoruces) even for not opened Session

Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=580524
Change-Id: Id24ec1c6032351a5600799363ec2f1d86723956a
Signed-off-by: Laurent Fasani <laurent.fasani@obeo.fr>
Signed-off-by: Glenn Plouhinec <glenn.plouhinec@obeo.fr>