[580691] Avoid duplicated tasks for Delete from Model without tool
This commit has no effect of added tests. They are already OK with the
previous commit. But as explained in the issue, if we observe in debug
the number of DeleteEObjectTask executed in the command, there are
potential "duplicates" (two DeleteEObjectTask that delete the same
DEdge). This commit resolves this "problem".

Bug: 580691
Change-Id: Ib7e4dcfcbf6cc3f800d242563bb05d639162c9c5
Signed-off-by: Maxime Porhel <maxime.porhel@obeo.fr>