[580673] Prevent loading option for aird in Git
Prevent loading option for aird with URI starting with a "commit" scheme
as it indicates the loading of an aird from Git. This happens when
comparing a current project with a previous version that is in Git.
Blocking the loading options prevent triggering the migration
participants on what should be a "read-only" aird as it is an old
version used for detecting changes here.

Bug: 580673
Change-Id: Ie5c2a7ced1d3aebdac3b07f3df18c3111e4379db
Signed-off-by: Steve Monnier <steve.monnier@obeo.fr>