[580801] Change tree expand behavior with right arrow key pressed
Before this commit, if the first expand of a tree item is done with the
keyboard, ie right arrow key pressed, the tree item is expanded, but
only one child without name is visible.
After this commit, the behavior for an expand done with the mouse is
similar that an expand done with the right arrow key pressed.

Bug: 580801
Cherry-picked-from: 549352
Change-Id: Ia56285159c0d429d5c8fbd2d305dafaed8af3b4d