[580786] Disable direct edit on key pressed for Tree and Table editor
This commit disables the direct edit "activation" on any alphanumeric
key pressed. By default, as the documentation always said, only F2 key,
Return key or double click enable the direct edit of a tree/table item.

As the behavior was wrong since the start, a property system has been
added to allow to retrieve the "wrong behavior". This property system is
"org.eclipse.sirius.ui.restoreBehaviorEnablingDirectEditOnAlphanumericKey"
and must be set to true to have the previous behavior.

Bug: 580786
Change-Id: If226676f1b673a287c81709c9d9aab7592a8f417