[580815] Execute closeOthersEditors in the same UI Thread.
When closing the EditingSession, the closeOthersEditors method was
executed asynchronously in the UI Thread, even if the current
closeEditor is called from the UI Thread. That might cause issues when
closing others editors referencing the session (editor reference after a
restart for instance): if the Sirius session has been closed in the
meantime, it cannot retrieve the session resources and cannot determine
which opened editor reference is related to the session.

Bug: 580815
Change-Id: Icb0ecde999b512a0522338528a0beb9999cdfe35