[581015] Support Collapse/Expand All with shift key pressed
Since bugzilla 549352 [1], in a tree editor, when the user clicks on the
right arrow key, the current tree item is expanded. If he clicks on the
left arrow key, the current tree item is collapsed.
With this commit, if the shift key is pressed with the arrow, an expand
all or collapse all is done.

[1] https://bugs.eclipse.org/bugs/show_bug.cgi?id=549352

Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=581015
Cherry-picked-from: https://bugs.eclipse.org/bugs/show_bug.cgi?id=580934
Change-Id: I12b1e88c9e71444105a11fa045b3fb4a0289f117