[580924] Fix notifications service after Reconnect or DragNDrop tool
This commit just handles the case of a "null" list of elements to select
later in the code, to be sure to enable fire notification in all cases.

All the changes are done in SelectDRepresentationElementsListener, but
it is to be sure to call enableFireNotification in sub class
DiagramSelectDRepresentationElementsListener , in method
setSelectionInAsyncExec(DialectEditor, List<DRepresentationElement>).

Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=580924
Change-Id: I21d54479b2c625e6dbfa8cb0a85ba0eba346824f