[581011] Move to Batik 1.16 and GMF Runtime 1.15.3
Batik 1.16 can only be consumed from a recent Orbit repo, which does
not include EasyMock anymore: the version we used was in the old
CVS-era part of Orbit, which is no longer provided starting from
2022-06.

This means than to get EasyMock we need to get it directly from Maven
Central. This is supported in recent versions of Maven and PDE
directly in .target files, but *not* (and probably never will) in the
TPD DSL.

As a consequence, the .target we need can not be generated anymore
using the TPD DSL, and we must drop its usage and write .target
directly from now on. This is not ideal but less of a hassle that it
used to be years ago when the TPD DSL was created.

Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=581011
Change-Id: I056fcfe412b25b59bced368e71292ec014525a7b
Signed-off-by: Pierre-Charles David <pierre-charles.david@obeo.fr>