[581232] Add tests to reveal the problem
Two tests have been added in this commit: one for pasting layout and one
for pasting format.
For information no test has been added for pasting style as the expected
result will be the same before and after the fix (no way to detect a
change in case of the "Copy style" action).

Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=581232
Change-Id: Id4984bb7880d823e0f190ebfcdcb8ad28db363c7